+++ app/controllers/likes_controller.rb
@@ -2,14 +2,17 @@ class LikesController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    like = Like.new(user_id: current_user.id, post_id: params[:post_id])
-    like.save
-    redirect_to post_path(params[:post_id])
+    unless current_user.like?(params[:post_id])
+      current_user.likes.create(post_id: params[:post_id])
+    end
+    redirect_back(fallback_location: root_path)
   end
 
   def destroy
-    like = Like.find_by(user_id: current_user.id, post_id: params[:post_id])
-    like.destroy
-    redirect_to post_path(params[:post_id])
+    if current_user.like?(params[:post_id])
+      like = current_user.likes.find_by(post_id: params[:post_id])
+      like.destroy
+    end
+    redirect_back(fallback_location: root_path)
   end
 end
